[code] improved test02
For decreasing systematic errors, test02 invalidates the positions
where the image will be written. So that every image has to b written
from new, befor test can pass.

The verification command does not return to the script if it fails,
console will be ended with returnvalue 1.

It's still a todo to control the position of the netX halted. It's not
possible to retrieve the value of the programmcounter into a tcl-variable.