[code] Added control of pc
Now the position of pc during the test is controlled too!
The test ow has two ways to fail and a single one to pass.

Failing by verify immediately without any huge error message and
failing with a hhuge printout if the pc was wrong. How ever, script
returnes 1 in case of failure.

A huge OK ist displayed followed by a picture in case the test has passed.